修复学分不能为小数的问题

+++ EzHerald/src/com/herald/ezherald/curriculum/CurriDBAdapter.java
@@ -184,7 +184,7 @@ public class CurriDBAdapter {
 			{
 				String name = mCursor.getString(mCursor.getColumnIndex(clCourseName));
 				String teacher = mCursor.getString(mCursor.getColumnIndex(clCourseLecturer));
-				int credit = mCursor.getInt(mCursor.getColumnIndex(clCourseCredit));
+				float credit = mCursor.getFloat(mCursor.getColumnIndex(clCourseCredit));
 				String weeks = mCursor.getString(mCursor.getColumnIndex(clCourseWeeks));
 				courseList.add(new Course(name,teacher,weeks,credit));
